requirements.txt
babel==2.16.0
blinker==1.9.0
click==8.1.8
docxcompose==1.4.0
docxtpl==0.19.0
et_xmlfile==2.0.0
Flask==3.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
openpyxl==3.1.5
pillow==11.0.0
python-docx==1.1.2
setuptools==75.6.0
six==1.17.0
ttkbootstrap==1.10.1
typing_extensions==4.12.2
Werkzeug==3.0.1
